// mini-erp-web/src/app/api/model/produtoResponseDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProdutoResponseDTO { 
    id?: number;
    nome?: string;
    preco?: number;
    quantidadeEstoque?: number;
    criadoEm?: string;
}