// mini-erp-web/src/app/api/model/produtoRequestDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProdutoRequestDTO { 
    nome: string;
    preco: number;
    quantidadeEstoque: number;
}